+++ tests/columns/test_datecolumn.py
@@ -16,45 +16,45 @@ import django_tables2 as tables
 
 def test_should_handle_explicit_format():
     class TestTable(tables.Table):
-        date = tables.DateColumn(format="D b Y")
+        date = tables.DateColumn(format='D b Y')
 
         class Meta:
-            default = "—"
+            default = '—'
 
-    table = TestTable([{"date": date(2012, 9, 11)},
-                       {"date": None}])
-    assert table.rows[0]["date"] == "Tue sep 2012"
-    assert table.rows[1]["date"] == "—"
+    table = TestTable([{'date': date(2012, 9, 11)},
+                       {'date': None}])
+    assert table.rows[0].get_cell('date') == 'Tue sep 2012'
+    assert table.rows[1].get_cell('date') == '—'
 
 
 def test_should_handle_long_format(settings):
-    settings.DATE_FORMAT = "D Y b"
+    settings.DATE_FORMAT = 'D Y b'
 
     class TestTable(tables.Table):
         date = tables.DateColumn(short=False)
 
         class Meta:
-            default = "—"
+            default = '—'
 
-    table = TestTable([{"date": date(2012, 9, 11)},
-                       {"date": None}])
-    assert table.rows[0]["date"] == "Tue 2012 sep"
-    assert table.rows[1]["date"] == "—"
+    table = TestTable([{'date': date(2012, 9, 11)},
+                       {'date': None}])
+    assert table.rows[0].get_cell('date') == 'Tue 2012 sep'
+    assert table.rows[1].get_cell('date') == '—'
 
 
 def test_should_handle_short_format(settings):
-    settings.SHORT_DATE_FORMAT = "b Y D"
+    settings.SHORT_DATE_FORMAT = 'b Y D'
 
     class TestTable(tables.Table):
         date = tables.DateColumn(short=True)
 
         class Meta:
-            default = "—"
+            default = '—'
 
-    table = TestTable([{"date": date(2012, 9, 11)},
-                       {"date": None}])
-    assert table.rows[0]["date"] == "sep 2012 Tue"
-    assert table.rows[1]["date"] == "—"
+    table = TestTable([{'date': date(2012, 9, 11)},
+                       {'date': None}])
+    assert table.rows[0].get_cell('date') == 'sep 2012 Tue'
+    assert table.rows[1].get_cell('date') == '—'
 
 
 def test_should_be_used_for_datefields():
@@ -68,4 +68,4 @@ def test_should_be_used_for_datefields():
         class Meta:
             model = DateModel
 
-    assert type(Table.base_columns["field"]) == tables.DateColumn
+    assert type(Table.base_columns['field']) == tables.DateColumn