+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Added Italian translation (#315 by [@paolodina](https://github.com/paolodina)
 - Added Dutch translation.
 - Fixed {% blocktrans %} template whitespace issues
+- Fixed errors when using a column named `items` (#316)
 
 ## v1.1.6 (2016-04-02)
 - Correct error message about request context processors for current Django (#314)